Add files via upload
“Advanced Hotel Room Booking System”. 
This system helps to manage bookings of hotel rooms.
Room Registration: In this part of the system, you can add rooms to the hotel. You need to add information like the unique ID of the room and the type of room. The types of rooms can be single, double, or deluxe.
Single Room: This room has a single bed, bathroom, TV, and closet. Price: $20 per day.
Double Room: This room has a double bed, bathroom, TV, and closet. Price: $35 per day.
Deluxe Room: This room has a minibar, a bathtub, a king-size bed, and a sitting area. Price: $55 per day.
Customer Registration: In this part of the system, the operator can add customers. They can put in details like their name, and contact email.
Room Booking: In this part of the system, customers can book a room. They can choose from the types of rooms that are available, based on the period they want to book. Customers should book a room for a specific period. If a room is booked for a specific period, it can’t be booked again by another customer for that period. The information about the booking should be saved both in the history of the room’s bookings and the customer’s bookings.
Bill Generation & Saving: In this part of the system, a detailed bill is made for each booking. The bill includes all the costs, like room charges, taxes (20%), and service fees (10% extra). The bill can be viewed in the console OR saved as a text file on the computer (file format: txt). The bill should have the room’s id & type, the customer’s name & contact email, the period of booking, the price for booking for the period, taxes(20%), service fee(10% above everything), and total amount.
Reports: You can make reports from the history of bookings of specific rooms and upcoming bookings.
Saving: The system should be able to save its whole state to a file system. This way, if you turn off the system and turn it on again, all the saved rooms can be read from the file system and available for the user to search.
The operator of the system should be able to do these tasks in the console:
Add a room
Add a customer
Book a room for the customer by the type of room they like and the period of time they choose (after successful booking, the bill should be shown in the console, and suggest the operator save the bill in the file system)
Save the state of the system in the file system by saying the full path of the file where it should be saved
Load the state of the system from the file system by saying the full path of the file where it was saved
Make a report for a specific room, which should have booking history (customer’s name and booking period in dates[start, end]), and save it in the file system by saying the full path of the file where it should be saved.